add interaction checks

+++ main.py
@@ -1,6 +1,8 @@
 import logging
 from os import getenv, name
 from urllib.parse import urlencode
+from base64 import urlsafe_b64encode, urlsafe_b64decode
+import pickle
 
 import disnake
 import uvloop
@@ -26,11 +28,16 @@ emojis = {
 }
 
 
-async def fetch_endpoint(url: str, param: dict = None) -> dict:
+async def fetch_endpoint(url: str, param: dict = {}) -> dict:
     param['agent'] = 'discord'
     async with ClientSession() as session:
         async with session.get(url + urlencode(param)) as response:
             return await response.json()
+
+async def message_hook(message, ID: str) -> None:
+    new_data = pickle.dumps({"channel_id": message.channel.id, "message_id": message.id}, protocol=pickle.HIGHEST_PROTOCOL)
+    await fetch_endpoint(url=f"{baseurl}set?", param={"ID": ID, "prefix": "message", "data": urlsafe_b64encode(new_data)})
+ 
 
 
 async def play_view(possible_moves: dict, score: int, name: str) -> disnake.ui.View:
@@ -121,36 +128,37 @@ class Bot(commands.AutoShardedBot):
 
 bot = Bot()
 
-
-@bot.slash_command()
-async def scores(inter: disnake.ApplicationCommandInteraction) -> None:
-    data = await fetch_endpoint(url=f"{baseurl}scores?")
-    await inter.response.send_message(
-        embed=disnake.Embed(description=data["discord"], color=0xCCCCCC)
-    )
-
-
 @bot.slash_command()
 async def play(inter: disnake.ApplicationCommandInteraction) -> None:
-    data = await fetch_endpoint(url=f"{baseurl}data?", param={"ID": str(inter.author.id), "name": inter.author.name})
-    print(data)
+    data = await fetch_endpoint(url=f"{baseurl}data?", param={"ID": inter.author.id, "name": inter.author.name})
     if data["can_continue"] == 1:
         view = continue_select_view()
     else:
         view = await play_view(possible_moves=data["possible_moves"], score=data['score'], name=inter.author.name)
-    await inter.response.send_message(file=disnake.File(data['image_path']), view=view)
+    msg = await inter.response.send_message(file=disnake.File(data['image_path']), view=view)
+    await message_hook(message=await inter.original_message(), ID=str(inter.author.id))
+
 
 @bot.event
 async def on_button_click(inter) -> None:
-    if inter.component.custom_id == "exit":
-        await inter.response.defer()
-        await inter.message.delete()
-        return
-    c = ''
-    data = await fetch_endpoint(url=f"{baseurl}move?", param={"ID": str(inter.author.id), "action": inter.component.custom_id})
-    if data["possible_moves"]['over']:
-        c = 'Game Over!'
     await inter.response.defer()
-    await inter.edit_original_message(content=c, file=disnake.File(data['image_path']), attachments=[], view=await play_view(possible_moves=data["possible_moves"], score=data['score'], name=inter.author.name))
+    raw = await fetch_endpoint(url=f"{baseurl}get?", param={"ID": inter.author.id, "name": inter.author.name, "prefix": "message"})
+    if raw['success']:
+        data = pickle.loads(urlsafe_b64decode(raw['data']))
+        msg = await inter.original_message()
+        if data['message_id'] == msg.id:
+            if inter.component.custom_id == "exit":
+                await inter.response.defer()
+                await inter.message.delete()
+                return
+            c = ''
+            data = await fetch_endpoint(url=f"{baseurl}move?", param={"ID": inter.author.id, "action": inter.component.custom_id})
+            if data["possible_moves"]['over']:
+                c = 'Game Over!'
+            await inter.edit_original_message(content=c, file=disnake.File(data['image_path']), attachments=[], view=await play_view(possible_moves=data["possible_moves"], score=data['score'], name=inter.author.name))
+        else:
+            pass
+    else:
+        pass
 
 bot.run(getenv("DISCORD_TOKEN"))
